requirements.txt
pandas>=2.1,<3
pyarrow>=15,<20
pydantic>=2.6,<3
pandera>=0.20,<0.26
duckdb>=0.10,<2
fpdf2>=2.7,<3
streamlit>=1.32,<2
altair>=5.2,<6
